requirements.txt
amqp==5.1.1
anyio==3.6.1
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.5
Django==4.1.2
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
dnspython==2.2.1
drf-yasg==1.21.4
email-validator==1.2.1
ez-setup==0.9
fastapi==0.79.0
filelock==3.7.1
gevent==21.12.0
greenlet==1.1.3
h11==0.13.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
phonenumberslite==8.12.57
Pillow==9.2.0
pipenv==2022.7.24
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-telegram-bot==13.13
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rabbitmq==0.2.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.19.1
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.2
vine==5.0.0
virtualenv==20.16.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
